+++ pkg/webhook/internal/cert/writer/certwriter.go
@@ -29,7 +29,7 @@ import (
 	admissionregistrationv1beta1 "k8s.io/api/admissionregistration/v1beta1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/cert/generator"
+	"sigs.k8s.io/controller-runtime/pkg/webhook/internal/cert/generator"
 )
 
 const (
