+++ pkg/webhook/internal/cert/writer/fs_test.go
@@ -30,8 +30,8 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/cert/generator"
-	fakegenerator "sigs.k8s.io/controller-runtime/pkg/webhook/cert/generator/fake"
+	"sigs.k8s.io/controller-runtime/pkg/webhook/internal/cert/generator"
+	fakegenerator "sigs.k8s.io/controller-runtime/pkg/webhook/internal/cert/generator/fake"
 )
 
 var _ = Describe("FSCertWriter", func() {
